ggml-zdnn: switch to common.hpp
